make eval more robust if pesq is not installed
